authorization route

+++ back/routes/treatments.route.js
@@ -1,10 +1,18 @@
 const express = require("express");
+const {Op} = require("sequelize")
 const treatments = express.Router();
 const Treatment = require("../models/Treatments");
+const {authToken } = require("../middlewares")
 
-treatments.get("/", async (req, res) => {
+treatments.use(authToken)
+
+treatments.get("/" ,async (req, res) => {
     try {
-        const treatments = await Treatment.findAll();
+        const treatments = await Treatment.findAll({
+            where : {
+                UserId : req.user.id,
+            }
+        });
         res.status(200).json(treatments);
     } catch (err) {
         res.status(400).send(err.message);
@@ -15,22 +23,22 @@ treatments.get("/:id", async (req, res) => {
     const { id } = req.params;
     console.log(id);
     try {
-        const Treatment = await Treatment.findAll({ where: { id } });
-        console.log(Treatment);
-        res.status(200).json(Treatment);
+        const treatment = await Treatment.findAll({ where: { id } });
+        res.status(200).json(treatment);
     } catch (err) {
         res.status(400).json(err);
     }
 });
 
 treatments.post("/", async (req, res) => {
-    const { name, duration } = req.body;
+    const { startDate, endDate } = req.body;
     try {
-        const Treatment = await Treatment.create({
-            name,
-            duration,
+        const treatment = await Treatment.create({
+            startDate,
+            endDate,
+            UserId: req.user.id
         });
-        res.status(201).json(Treatment);
+        res.status(201).json(treatment);
     } catch (err) {
         res.status(422).json(err);
     }
@@ -38,10 +46,10 @@ treatments.post("/", async (req, res) => {
 
 treatments.put("/:id", async (req, res) => {
     const { id } = req.params;
-    const { name, duration } = req.body;
+    const { startDate, endDate } = req.body;
     try {
         const Treatment = await Treatment.update(
-            { name, duration },
+            { startDate, endDate },
             { where: { id } }
         );
         res.status(202).send("Utilisateur modifié");
@@ -53,7 +61,7 @@ treatments.put("/:id", async (req, res) => {
 treatments.delete("/:id", async (req, res) => {
     const { id } = req.params;
     try {
-        const Treatment = await Treatment.destroy({ where: { id } });
+        const Treatment = await Treatment.destroy({ where: { id, UserId: req.user.id } });
         res.status(205).send("L'utilisateur a bien été supprimé");
     } catch (err) {
         res.status(422).json(err);
